+++ src/lib.rs
@@ -1,4 +1,5 @@
 mod address;
+mod callback;
 mod consensus;
 mod crypto;
 mod macros;
@@ -106,6 +107,7 @@ fn kaspa(py: Python, m: &Bound<'_, PyModule>) -> PyResult<()> {
     m.add_class::<wallet::core::utxo::balance::PyBalance>()?;
     m.add_class::<wallet::core::utxo::balance::PyBalanceStrings>()?;
     m.add_class::<wallet::core::utxo::context::PyUtxoContext>()?;
+    m.add_class::<wallet::core::utxo::processor::PyUtxoProcessorEvent>()?;
     m.add_class::<wallet::core::utxo::processor::PyUtxoProcessor>()?;
 
     m.add_function(wrap_pyfunction!(